+++ Sources/AuvasaKit/Services/Static/GTFSImporter.swift
@@ -35,7 +35,7 @@ public actor GTFSImporter {
     private let gtfsURL: URL
 
     public static let defaultGTFSURL: URL = {
-        guard let url = URL(string: "https://www.auvasa.es/wp-file-download/datos-gtfs/") else {
+        guard let url = URL(string: "http://212.170.201.204:50080/GTFSRTapi/api/GTFSFile") else {
             fatalError("Invalid GTFS URL")
         }
         return url